@@ -17,6 +17,7 @@ dependencies:
   intl: ^0.17.0
   json_annotation: ^4.0.0
   provider: ^5.0.0
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
   build_runner:
@@ -28,7 +29,7 @@ dev_dependencies:
   test: ^1.16.5
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 flutter_icons:
   android: launcher_icon
