@@ -3,7 +3,6 @@ import 'dart:async';
 import 'package:bloc/bloc.dart';
 import 'package:catan_master/domain/feedback/feedback_message.dart';
 import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
 
 part 'feedback_event.dart';
 part 'feedback_state.dart';
@@ -23,11 +22,11 @@ class FeedbackBloc extends Bloc<FeedbackEvent, FeedbackState> {
     add(FeedbackEvent(feedbackMessage));
   }  
   
-  snackbar(String message, {Severity severity}) {
+  snackbar(String message, {Severity severity = Severity.message}) {
     add(FeedbackEvent(FeedbackMessage.snackbar(message, severity: severity)));
   }
   
-  dialog(String message, {@required String title, Severity severity}) {
+  dialog(String message, {required String title, Severity severity = Severity.message}) {
     add(FeedbackEvent(FeedbackMessage.dialog(message, title: title, severity: severity)));
   }
 }
