@@ -3,28 +3,28 @@ part of 'players_bloc.dart';
 @immutable
 abstract class PlayerState {
 
-  final Message message;
+  final Message? message;
 
   PlayerState({this.message});
 
-  PlayerState copyWith({Message message});
+  PlayerState copyWith({Message? message});
 
 }
 
 class InitialPlayersState extends PlayerState {
 
-  InitialPlayersState({Message message}) : super(message: message);
+  InitialPlayersState({Message? message}) : super(message: message);
 
-  InitialPlayersState copyWith({Message message}) {
+  InitialPlayersState copyWith({Message? message}) {
     return InitialPlayersState(message: message ?? message);
   }
 }
 
 class PlayersLoading extends PlayerState {
 
-  PlayersLoading({Message message}) : super(message: message);
+  PlayersLoading({Message? message}) : super(message: message);
 
-  PlayersLoading copyWith({Message message}) {
+  PlayersLoading copyWith({Message? message}) {
     return PlayersLoading(message: message ?? message);
   }
 
@@ -34,23 +34,23 @@ class PlayersLoaded extends PlayerState {
 
   final List<Player> players;
 
-  PlayersLoaded(List<Player> players, {Message message}) :
+  PlayersLoaded(List<Player?> players, {Message? message}) :
         this.players = List.unmodifiable(players),
         super(message: message);
 
-  PlayersLoaded copyWith({Player newPlayer, Message message}) {
+  PlayersLoaded copyWith({Player? newPlayer, Message? message}) {
     return PlayerAdded(players, newPlayer: newPlayer, message: message ?? message);
   }
 
-  Player getPlayer(String username) => players.firstWhere((p) => p.username == username, orElse: () => null);
+  Player? getPlayer(String username) => players.firstWhereOrNull((p) => p.username == username);
 
 }
 
 class PlayerAdded extends PlayersLoaded {
 
-  final Player newPlayer;
+  final Player? newPlayer;
 
-  PlayerAdded(List<Player> players, {this.newPlayer, Message message}) : super(
+  PlayerAdded(List<Player> players, {this.newPlayer, Message? message}) : super(
       List.from(players)..add(newPlayer),
       message: message
   );
@@ -59,7 +59,7 @@ class PlayerAdded extends PlayersLoaded {
 class PlayerEdited extends PlayersLoaded {
 
   //TODO move to Players object
-  final Player editedPlayer;
+  final Player? editedPlayer;
 
   PlayerEdited(List<Player> players, {this.editedPlayer}) : super(List.from(players)..add(editedPlayer));
 
