@@ -1,12 +1,10 @@
-import 'package:meta/meta.dart';
-
 class Message {
 
   final String message;
   final MessageType type;
-  final String extraInfo;
+  final String? extraInfo;
 
-  Message({@required this.message, this.type = MessageType.error, this.extraInfo});
+  Message({required this.message, this.type = MessageType.error, this.extraInfo});
 
   Message.error(this.message, {this.extraInfo}) : this.type = MessageType.error;
   Message.warning(this.message, {this.extraInfo}) : this.type = MessageType.warning;
