@@ -14,17 +14,14 @@ class Player extends Equatable {
   final Color color;
 
   Player({
-    @required String username,
-    @required this.name,
-    @required this.gender,
-    @required Color color
+    required String username,
+    required this.name,
+    required this.gender,
+    required Color color
   }) :
-        assert(username != null && username.isNotEmpty),
+        assert(username.isNotEmpty),
         this.username = username.toLowerCase(),
-        assert(name != null),
         assert(name.isNotEmpty),
-        assert(gender != null),
-        assert(color != null),
         assert(color.alpha == 255),
         this.color = Color(color.value);
 
@@ -42,21 +39,21 @@ class PlayerStatistics extends Equatable {
 
   final Player player;
   final List<Game> games;
-  final Player bestBuddy;
+  final Player? bestBuddy;
   //TODO include 'Base' as an option here
-  final CatanExpansion mostPlayedExpansion;
-  final CatanExpansion mostWonExpansion;
+  final CatanExpansion? mostPlayedExpansion;
+  final CatanExpansion? mostWonExpansion;
   final int rank;
   final List<Prize> prizes;
 
   PlayerStatistics._({
-    @required this.player,
-    @required this.games,
-    @required this.bestBuddy,
-    @required this.mostPlayedExpansion,
-    @required this.mostWonExpansion,
-    @required this.rank,
-    @required this.prizes,
+    required this.player,
+    required this.games,
+    required this.bestBuddy,
+    required this.mostPlayedExpansion,
+    required this.mostWonExpansion,
+    required this.rank,
+    required this.prizes,
   });
 
   factory PlayerStatistics.fromGames(Player player, Games allGames) {
@@ -66,7 +63,7 @@ class PlayerStatistics extends Equatable {
     Map<Player, int> gamesWon = {};
 
     Map<CatanExpansion, int> expansionsPlayed = {};
-    Map<CatanExpansion, int> expansionsWon = {};
+    Map<CatanExpansion?, int> expansionsWon = {};
 
     games.forEach((game) {
       game.players.forEach((p) => gamesPlayed[p] = (gamesPlayed[p] ?? 0)+1);
@@ -101,7 +98,7 @@ class PlayerStatistics extends Equatable {
 
   int getNbWinsOutOfLast(int nb) => games.take(nb).where((g) => g.winner == player).length;
 
-  Game get lastGame => games.firstOrNull;
+  Game? get lastGame => games.firstOrNull;
 
   List<bool> getWinOrLose(int nb) => games.take(nb).toList().map((g) => g.winner == player).toList();
 
@@ -112,7 +109,7 @@ class PlayerStatistics extends Equatable {
 
 extension HighestValue<K> on Map<K, int> {
 
-  K get highestValue {
+  K? get highestValue {
     if (isEmpty) return null;
     return entries.toList().reduce((v1, v2) => v1.value >= v2.value ? v1 : v2).key;
   }
