@@ -58,7 +58,7 @@ class CatanMasterApp extends StatelessWidget {
             } else if (settings.name == '/players/detail') {
               return MaterialPageRoute(
                 builder: (context) {
-                  var player = (settings.arguments as Map)["player"] as Player;
+                  var player = (settings.arguments as Map)["player"] as Player?;
                   if (player == null) {
                     return Text("Error: no player provided");
                   }
@@ -79,7 +79,7 @@ class CatanMasterApp extends StatelessWidget {
 class CatanMasterHomeScreen extends StatelessWidget {
 
   final HomePageTab tab;
-  final Function onAddTapped;
+  final Function? onAddTapped;
 
   CatanMasterHomeScreen({this.tab = HomePageTab.games, this.onAddTapped});
 
@@ -113,7 +113,7 @@ class CatanMasterHomeScreen extends StatelessWidget {
             onPressed: () {
               switch(tab) {
                 case HomePageTab.games:
-                  var playersState = BlocProvider.of<PlayersBloc>(context).state;
+                  PlayerState playersState = BlocProvider.of<PlayersBloc>(context).state;
                   if (playersState is PlayersLoaded) {
                     if (playersState.players.length > 1) {
                       Navigator.of(context).pushNamed("/games/add");
@@ -152,8 +152,8 @@ class CatanMasterHomeScreen extends StatelessWidget {
 class CatanFloatingActionButton extends StatelessWidget {
 
   const CatanFloatingActionButton({
-    Key key,
-    @required this.onPressed,
+    Key? key,
+    required this.onPressed,
   }) : super(key: key);
 
   final Function() onPressed;
@@ -163,7 +163,7 @@ class CatanFloatingActionButton extends StatelessWidget {
     return Hero(
       tag: "catan_fab",
       flightShuttleBuilder: (flightContext, animation, flightDirection, fromHeroContext, toHeroContext) {
-        final Hero toHero = toHeroContext.widget;
+        final Hero toHero = toHeroContext.widget as Hero;
         return RotationTransition(
           turns: Tween<double>(begin: 0.0, end: 1.0)
               .animate(CurvedAnimation(parent: animation, curve: Curves.easeInOutSine)),
