@@ -7,13 +7,12 @@ import 'package:catan_master/domain/games/game_repository.dart';
 import 'package:catan_master/domain/players/player_repository.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:meta/meta.dart';
 
 class CatanMasterBlocProvider extends StatelessWidget {
 
   final Widget child;
 
-  CatanMasterBlocProvider({Key key, @required this.child}) : super(key: key);
+  CatanMasterBlocProvider({Key? key, required this.child}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
