@@ -4,10 +4,10 @@ import 'package:flutter/material.dart';
 class EmptyListMessage extends StatelessWidget {
 
   final Widget title;
-  final Widget subtitle;
-  final Widget action;
+  final Widget? subtitle;
+  final Widget? action;
 
-  EmptyListMessage({@required this.title, this.subtitle, this.action}) : assert(title != null);
+  EmptyListMessage({required this.title, this.subtitle, this.action});
 
   @override
   Widget build(BuildContext context) {
@@ -17,15 +17,15 @@ class EmptyListMessage extends StatelessWidget {
         mainAxisSize: MainAxisSize.max,
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
-          DefaultTextStyle(style: Theme.of(context).textTheme.headline5, textAlign: TextAlign.center, child: title),
+          DefaultTextStyle(style: Theme.of(context).textTheme.headline5!, textAlign: TextAlign.center, child: title),
           if (subtitle != null) SizedBox(height: 8.0,),
           if (subtitle != null) DefaultTextStyle(
-              style: Theme.of(context).textTheme.bodyText2,
+              style: Theme.of(context).textTheme.bodyText2!,
               textAlign: TextAlign.center,
-              child: subtitle
+              child: subtitle!
           ),
           if (action != null) SizedBox(height: 8.0,),
-          if (action != null) action,
+          if (action != null) action!,
         ],
       ),
     );
