@@ -21,17 +21,17 @@ class GamesLoaded extends GamesState {
 
 class GameAdded extends GamesLoaded {
 
-  final Game/*!*/ newGame;
+  final Game newGame;
 
-  GameAdded(Games games, {this.newGame}) : super(Games(List.from(games.games)..add(newGame)));
+  GameAdded(Games games, {required this.newGame}) : super(Games(List.from(games.games)..add(newGame)));
 
 }
 
 class GameEdited extends GamesLoaded {
 
   //TODO move to games
-  final Game/*!*/ editedGame;
+  final Game editedGame;
 
-  GameEdited(Games games, {this.editedGame}) : super(Games(List.from(games.games)..add(editedGame)));
+  GameEdited(Games games, {required this.editedGame}) : super(Games(List.from(games.games)..add(editedGame)));
 
 }
