@@ -1,5 +1,5 @@
 extension ListExtensions<T> on Iterable<T> {
 
-  T get firstOrNull => isEmpty ? null : first;
+  T? get firstOrNull => isEmpty ? null : first;
 
 }
