@@ -16,6 +16,6 @@ abstract class GameRepository {
 
   Future<Either<Failure, void>> deleteGame(Game game);
 
-  Future<Either<Failure, void>> undoDelete({Game game});
+  Future<Either<Failure, void>> undoDelete({Game? game});
 
 }
