@@ -17,8 +17,9 @@ abstract class PlayerRepository {
 /// When you try to delete a player which still has games assigned to him
 class PlayerHasGamesFailure extends Failure {
 
-  PlayerHasGamesFailure(Player player, {int nbGames}) : super(
-      "Player '$player' has ${nbGames.toString() + " " ?? ""}game(s).\nOnly players with no games can be deleted. Either delete all the games or remove '$player' from those games."
+  PlayerHasGamesFailure(Player player, {required int nbGames}) : super(
+      "Player '$player' has $nbGames game(s).\nOnly players with no games can be deleted. "
+          "Either delete all the games or remove '$player' from those games."
   );
 
 }
