@@ -31,7 +31,7 @@ class CatanIcons {
   CatanIcons._();
 
   static const _kFontFam = 'Catan';
-  static const _kFontPkg = null;
+  static const dynamic _kFontPkg = null;
 
   static const IconData dice = IconData(0xe800, fontFamily: _kFontFam, fontPackage: _kFontPkg);
   static const IconData trophy = IconData(0xe801, fontFamily: _kFontFam, fontPackage: _kFontPkg);
